Use NumPy recommended approach to out-of-bound conversion

+++ aesara/misc/safe_asarray.py
@@ -32,7 +32,7 @@ def _asarray(a, dtype, order=None):
     if str(dtype) == "floatX":
         dtype = config.floatX
     dtype = np.dtype(dtype)  # Convert into dtype object.
-    rval = np.asarray(a, dtype=dtype, order=order)
+    rval = np.asarray(a, order=order).astype(dtype)
     # Note that dtype comparison must be done by comparing their `num`
     # attribute. One cannot assume that two identical data types are pointers
     # towards the same object (e.g. under Windows this appears not to be the
